Updated some stuff

+++ app/src/main/java/com/acg/outtamycircle/ClientScreen.java
@@ -63,11 +63,11 @@ public class ClientScreen extends ClientServerScreen {
             switch (interpreter.getType(message)){
                 case MOVE_SERVER: {
                     int objectId = interpreter.getObjectId(message);
-                    short posX = interpreter.getPosX(message);
-                    short posY = interpreter.getPosY(message);
+                    float posX = interpreter.getPosX(message);
+                    float posY = interpreter.getPosY(message);
                     float rotation = interpreter.getRotation(message);
                     DrawableComponent comp = (DrawableComponent) status.characters[objectId].getComponent(Component.Type.Drawable);
-                    comp.setX(posX).setY(posY);
+                    comp.setX((int)posX).setY((int)posY);
                 }
                 break;
                 case ATTACK: {
